Main design finished
I have finished the all main design. At this moment we only have the design of the dropdown finished. In this code the display option is disable so we can see the dropdown option. We dont have more options now but i will add some in the future once this one is completed.